Update NAD.md
Hi Rob
Ntopng hasn't got great visibility in the ICS world, but the tool gives great network insights, is a one in all solution (easy to install, monitoring-logging-alarming included) and as open source project free of charge. Of course it does not have the feature set like Dragos or Nozomi, but for small or medium sized companies it can be a good fit. Currently only the IEC 60870-5-104 protocol is fully implemented, but we are constantly expanding.
Thank you, Martin